fix refactoring errors from base branch
